[refactor] form component 변경 #
